Factored out data sanitizer logic
