+++ apps/stats/src/views/Stats/Web/components/SourcesCard.tsx
@@ -1,34 +1,9 @@
 import React from 'react';
+import {BaseSourceData, ProcessedSourceData, extendSourcesWithPercentages, processSources} from '@tryghost/admin-x-framework';
 import {Button, Card, CardContent, CardDescription, CardFooter, CardHeader, CardTitle, DataList, DataListBar, DataListBody, DataListHead, DataListHeader, DataListItemContent, DataListItemValue, DataListItemValueAbs, DataListItemValuePerc, DataListRow, LucideIcon, Separator, Sheet, SheetContent, SheetDescription, SheetHeader, SheetTitle, SheetTrigger, Table, TableBody, TableCell, TableHead, TableHeader, TableRow, formatNumber, formatPercentage} from '@tryghost/shade';
-import {getFaviconDomain} from '@tryghost/admin-x-framework';
 
 // Default source icon URL - apps can override this
 const DEFAULT_SOURCE_ICON_URL = 'https://www.google.com/s2/favicons?domain=ghost.org&sz=64';
-
-// Base interface for all source data types
-interface BaseSourceData {
-    source?: string | number;
-    visits?: number;
-    free_members?: number;
-    paid_members?: number;
-    mrr?: number;
-    [key: string]: unknown;
-}
-
-// Processed source data with pre-computed display values
-interface ProcessedSourceData {
-    source: string;
-    visits: number;
-    isDirectTraffic: boolean;
-    iconSrc: string;
-    displayName: string;
-    linkUrl?: string;
-    percentage?: number;
-    // Additional fields for growth data
-    free_members?: number;
-    paid_members?: number;
-    mrr?: number;
-}
 
 interface SourcesTableProps {
     data: ProcessedSourceData[] | null;
@@ -143,7 +118,7 @@ const SourcesTable: React.FC<SourcesTableProps> = ({data, mode, defaultSourceIco
     );
 };
 
-interface SourcesProps {
+interface SourcesCardProps {
     title?: string;
     description?: string;
     data: BaseSourceData[] | null;
@@ -156,7 +131,7 @@ interface SourcesProps {
     getPeriodText?: (range: number) => string;
 }
 
-export const Sources: React.FC<SourcesProps> = ({
+export const SourcesCard: React.FC<SourcesCardProps> = ({
     title = 'Top Sources',
     description,
     data,
@@ -170,119 +145,22 @@ export const Sources: React.FC<SourcesProps> = ({
 }) => {
     // Process and group sources data with pre-computed icons and display values
     const processedData = React.useMemo(() => {
-        if (!data) {
-            return [];
-        }
-
-        const sourceMap = new Map<string, ProcessedSourceData>();
-        let directTrafficTotal = mode === 'visits' ? 0 : undefined;
-        const directTrafficData = mode === 'growth' ? {
-            free_members: 0,
-            paid_members: 0,
-            mrr: 0
-        } : undefined;
-
-        // Process each source and group direct traffic
-        data.forEach((item) => {
-            const {domain: faviconDomain, isDirectTraffic} = getFaviconDomain(item.source, siteUrl);
-            const visits = Number(item.visits || 0);
-
-            if (isDirectTraffic || !item.source || item.source === '') {
-                // Accumulate all direct traffic
-                if (mode === 'visits') {
-                    directTrafficTotal! += visits;
-                } else if (mode === 'growth' && directTrafficData) {
-                    directTrafficData.free_members += item.free_members || 0;
-                    directTrafficData.paid_members += item.paid_members || 0;
-                    directTrafficData.mrr += item.mrr || 0;
-                }
-            } else {
-                // Keep other sources as-is
-                const sourceKey = String(item.source);
-                const iconSrc = faviconDomain
-                    ? `https://www.faviconextractor.com/favicon/${faviconDomain}?larger=true`
-                    : defaultSourceIconUrl;
-                const linkUrl = faviconDomain ? `https://${faviconDomain}` : undefined;
-
-                if (sourceMap.has(sourceKey)) {
-                    const existing = sourceMap.get(sourceKey)!;
-                    existing.visits += visits;
-                    if (mode === 'growth') {
-                        existing.free_members = (existing.free_members || 0) + (item.free_members || 0);
-                        existing.paid_members = (existing.paid_members || 0) + (item.paid_members || 0);
-                        existing.mrr = (existing.mrr || 0) + (item.mrr || 0);
-                    }
-                } else {
-                    const processedItem: ProcessedSourceData = {
-                        source: sourceKey,
-                        visits,
-                        isDirectTraffic: false,
-                        iconSrc,
-                        displayName: sourceKey,
-                        linkUrl
-                    };
-
-                    if (mode === 'growth') {
-                        processedItem.free_members = item.free_members || 0;
-                        processedItem.paid_members = item.paid_members || 0;
-                        processedItem.mrr = item.mrr || 0;
-                    }
-
-                    sourceMap.set(sourceKey, processedItem);
-                }
-            }
+        return processSources({
+            data,
+            mode,
+            siteUrl,
+            siteIcon,
+            defaultSourceIconUrl
         });
-
-        // Add consolidated direct traffic entry if there's any
-        const hasDirectTraffic = mode === 'visits'
-            ? directTrafficTotal! > 0
-            : directTrafficData && (directTrafficData.free_members > 0 || directTrafficData.paid_members > 0 || directTrafficData.mrr > 0);
-
-        if (hasDirectTraffic) {
-            const directEntry: ProcessedSourceData = {
-                source: 'Direct',
-                visits: mode === 'visits' ? directTrafficTotal! : 0,
-                isDirectTraffic: true,
-                iconSrc: siteIcon || defaultSourceIconUrl,
-                displayName: 'Direct',
-                linkUrl: undefined
-            };
-
-            if (mode === 'growth' && directTrafficData) {
-                directEntry.free_members = directTrafficData.free_members;
-                directEntry.paid_members = directTrafficData.paid_members;
-                directEntry.mrr = directTrafficData.mrr;
-            }
-
-            sourceMap.set('Direct', directEntry);
-        }
-
-        // Convert back to array and sort
-        const result = Array.from(sourceMap.values());
-
-        if (mode === 'growth') {
-            // Sort by total impact (prioritizing MRR, then paid members, then free members)
-            return result.sort((a, b) => {
-                const aScore = (a.mrr || 0) * 100 + (a.paid_members || 0) * 10 + (a.free_members || 0);
-                const bScore = (b.mrr || 0) * 100 + (b.paid_members || 0) * 10 + (b.free_members || 0);
-                return bScore - aScore;
-            });
-        } else {
-            // Sort by visits
-            return result.sort((a, b) => b.visits - a.visits);
-        }
     }, [data, siteUrl, siteIcon, mode, defaultSourceIconUrl]);
 
     // Extend processed data with percentage values for visits mode
     const extendedData = React.useMemo(() => {
-        if (mode === 'growth') {
-            return processedData;
-        }
-
-        return processedData.map(item => ({
-            ...item,
-            percentage: totalVisitors > 0 ? (item.visits / totalVisitors) : 0
-        }));
+        return extendSourcesWithPercentages({
+            processedData,
+            totalVisitors,
+            mode
+        });
     }, [processedData, totalVisitors, mode]);
 
     const topSources = extendedData.slice(0, 10);
@@ -340,4 +218,4 @@ export const Sources: React.FC<SourcesProps> = ({
     );
 };
 
-export default Sources;
+export default SourcesCard;
